model/endpoint/requirements.txt
transformers>=4.56
Pillow
ultralytics
